Extract form types to own file
