+++ gallery/src/demos/demo-ha-form.ts
@@ -1,10 +1,11 @@
 /* eslint-disable lit/no-template-arrow */
 import { LitElement, TemplateResult, css, html } from "lit";
 import { customElement } from "lit/decorators";
-import { computeInitialData } from "../../../src/components/ha-form/ha-form";
+import { computeInitialHaFormData } from "../../../src/components/ha-form/compute-initial-ha-form-data";
 import "../../../src/components/ha-card";
 import { applyThemesOnElement } from "../../../src/common/dom/apply_themes_on_element";
-import type { HaFormSchema } from "../../../src/components/ha-form/ha-form";
+import type { HaFormSchema } from "../../../src/components/ha-form/types";
+import "../../../src/components/ha-form/ha-form";
 
 const SCHEMAS: {
   title: string;
@@ -188,7 +189,7 @@ const SCHEMAS: {
 
 @customElement("demo-ha-form")
 class DemoHaForm extends LitElement {
-  private data = SCHEMAS.map(({ schema }) => computeInitialData(schema));
+  private data = SCHEMAS.map(({ schema }) => computeInitialHaFormData(schema));
 
   protected render(): TemplateResult {
     return html`
