+++ src/components/ha-form/compute-initial-ha-form-data.ts
@@ -1,0 +1,29 @@
+import { HaFormSchema } from "./types";
+
+export const computeInitialHaFormData = (
+  schema: HaFormSchema[]
+): Record<string, any> => {
+  const data = {};
+  schema.forEach((field) => {
+    if (field.description?.suggested_value) {
+      data[field.name] = field.description.suggested_value;
+    } else if ("default" in field) {
+      data[field.name] = field.default;
+    } else if (!field.required) {
+      // Do nothing.
+    } else if (field.type === "boolean") {
+      data[field.name] = false;
+    } else if (field.type === "string") {
+      data[field.name] = "";
+    } else if (field.type === "integer") {
+      data[field.name] = "valueMin" in field ? field.valueMin : 0;
+    } else if (field.type === "constant") {
+      data[field.name] = field.value;
+    } else if (field.type === "float") {
+      data[field.name] = 0.0;
+    } else if (field.type === "select" && field.options.length) {
+      data[field.name] = field.options[0][0];
+    }
+  });
+  return data;
+};
