+++ src/data/data_entry_flow.ts
@@ -1,5 +1,5 @@
 import { Connection } from "home-assistant-js-websocket";
-import { HaFormSchema } from "../components/ha-form/ha-form";
+import type { HaFormSchema } from "../components/ha-form/types";
 import { ConfigEntry } from "./config_entries";
 
 export interface DataEntryFlowProgressedEvent {
