+++ src/dialogs/config-flow/show-dialog-data-entry-flow.ts
@@ -1,6 +1,6 @@
 import { TemplateResult } from "lit";
 import { fireEvent } from "../../common/dom/fire_event";
-import { HaFormSchema } from "../../components/ha-form/ha-form";
+import type { HaFormSchema } from "../../components/ha-form/types";
 import {
   DataEntryFlowStep,
   DataEntryFlowStepAbort,
